Refactor client: split commands and messaging
