+++ bitchat/__init__.py
@@ -1,0 +1,2 @@
+from .client import BitchatClient, main
+__all__ = ["BitchatClient", "main"]
