+++ bitchat/compression.py
