+++ requirements.txt
@@ -1,0 +1,5 @@
+bleak
+aioconsole
+pybloom-live
+cryptography
+lz4
